Use go code to start defra inside the indexer process instead of using 'make bootstrap' and starting defra in a separate process. This is much more lightweight and less error prone.
